Test: Export Test Output From Containers
Previously, if running the test cases inside of a container (with e.g.
--runtime=docker), the output would go to a directory inside of the
container. Now, we bind-mount in a directory from outside, so that the
container can write to the standard location (currently, the directory
"test_output" in the project root). We still allow for users to manually
specify an output directory, which should work now.