Add 'getfirst' to gettrav
In the SFS and BFS traversal methods, we need to get the "first"
sub-candidate for a given position; this was previously done by
creating a new TravPos object and calling newtrav. Now, a 'getfirst'
option has been added to gettrav so we can use that instead.

Should provide some speed-up, I hope.