RndLearner: Add Saving Lemmas In-Between Runs
Now, when the program exits (cleanly), the current learned lemmas will
be serialized to a file; this file is re-read on subsequent runs (on the
same benchmark), and the lemmas within are added to the known set of
lemmas. If the conjunction of lemmas is safe, the program exits.

Note that this means that subsequent runs of the program on the same
benchmark will just read the lemmas from the file and exit immediately;
delete/rename the file if this is undesired.