RndLearner: Adaptively Strengthen Candidate
If a candidate fails the inductive check, we then conjoin it with
the post condition; if this succeeds the inductive check (as it
sometimes does) then we use that as the candidate. This does not seem to
significantly increase the solve time, but does solve several benchmarks
in many fewer iterations.

Note this only works for V1; V2 and V3 support later.