Better V1 Adaptive Strengthening
Now, we try the strengthened version first, instead of only if the
non-strengthened version wasn't a lemma. This results in the same number
of iterations for the adaptive approach and the purely-strengthened
approach (i.e. the one that V2 and V3 take currently).

When combined with weakening, i.e. --inv-templ 3, we can solve several
benchmarks in a single iteration (which is basically what bootstrapping
would be doing if present in V1).