Grammars: First Stage of Refactoring
Large refactor: Introduces separate objects for traversals, constraints,
and grammars. Most helper functions moved out of GramFac.hpp (e.g.
collapsePT is now in ParseTree.hpp as a conversion to Expr); code
dealing with Grammar I/O is in CFGUtils to avoid cluttering Grammar.hpp
(to be reconsidered). 'Grammar' object is intended to be an easy way to
create grammars on the fly; most methods unimplemented, existing code is
just internally accessing the private members. I'll be migrating code
over to the new interface in the future commit(s).