Update to Latest Aeval Interface
The version previously used had a somewhat different interface in terms
of constructors and function headers; this updates the code to use the
new interface.

Note that Aeval supports a '--debug' option now; I'm currently
statically passing 0 for now printing, but passing 'printLog' so that
the debug output matches that of Freqhorn overall is also possible.