Grammars: Second Stage of Refactoring
The last commit introduced the Grammar object, but didn't actually make
use of any of its methods (just using it as a struct). Now the code
makes use of the methods, i.e. the parsing code creates it as any other
user would (with e.g. addProd()) and the traversal code uses the public
properties. Some regressions may be present, but all test cases are
currently passing.

Only known issue is when priorities are present in the grammar, but
that's been broken for a few commits now.