Testing Script Refactor
Not strictly necessary, but will enable the creation of an automated
benchmarking process (with nice CSV output instead of the blob we
typically deal with) in the near future. The test-specific code has been
split from the job scheduling code, so that the hypothetical
benchmarking script can reuse that code.

Also, some '#ifndef' guards for the common shell script code.