Better Recursive Depth Calculation
Previously, a series of productions like 'A -> B B; B -> A' would not
have been correctly identified as recursive. This isn't a problem for
the grammars we'd previously been dealing with, but I've now encountered
some like that. Now we calculate the transitive connectedness of each
non-terminal and use that to determine recursion (recursive if A is
transitively connected to A). This is cached for speed.

Also, gdb_expr.py now correctly prints MPZ 0s.