Naive Iterative Deepening
Where "Iterative Deepening" means check all candidates with recursive
depth 0 first, then 1, then etc. up to '--maxrecdepth'.

Tool option is '--iter-deepen'

"Naive" in the sense that it's quite wasteful. The coroutine traversal
at least tries to be smart in which cache entries it discards (so we at
least don't have to run the coroutines for the old candidates) but
NewTrav does nothing. We make no attempt (other than the facility
already present in GramEnum) to avoid candidate duplication.

Also, adds some test cases for the option.

Note: I have code deep in my 'git stash' for smarter iterative deepening
in NewTrav, which likely will be integrated in a later commit.