Add Support For Enumerative Synthesis
For functions with a grammar provided, we now support enumerating
candidates directly from the grammar. This is done in a fairly naive,
brute-force manner, just like my (not committed) enumeration support in
FreqHorn.

Use with '--sygus-mode enum'; an automated decider between the two
techniques will come later.

Realizability is assumed; unrealizability detection might be later.
I.e. if a candidate of the grammar meets the specification, then the
solver will eventually terminate. Otherwise, if the grammar is infinite
it will never terminate. If the grammar is finite, it will terminate.