Fix Segfaults With CFGUtils' Caches
The caches used to live statically, which was a problem because they'd
hold on to Expr's (which have dumb pointers to an ExprFactory which
exists on the stack). Thus, when the ExprFactory was deallocated (from a
return in its stack frame) there'd be a dangling pointer.

Now, there's an (empty) CFGUtils object with a reference-counted
singleton cache. If there are no users of the cache, it gets
deallocated. Each user of CFGUtils owns one of these objects (which can
be constructed from nothing unlike SMTUtils), thus the ownership of the
cache effectively becomes the same as that of the ExprFactory.