Lazy TravPos Children/Queue Items
In short, we make some attempt to determine when children/queue items
won't be needed any more and if so, we don't keep track of them.
Currently, this means we 'pop' from the children for NTs and 'pop' from
the queue when those items are done. Additionally, we defer some memory
allocations to the time they're actually needed.

These both cut memory usage by a fair amount (about %25), though further
analysis may show that more pruning is possible.