Further TravPos Memory Optimizations
Compilers typically add padding bytes in structs to align them with the
CPU's word size, on x86(_64) this is 4 bytes. This means that my
'std::pair<bool,TravPos*>' was actually 16 bytes instead of the ideal 9.
To fix this, we just encode the bool in the top bit of the pointer,
which I have always found unused.

Additionally, CircularInt's 'limit' was redundant, since we know it's
either the size of 'children' or of 'queue', which is already being
stored. Fixing this requires embedding the same logic into TravPos.
This has the nice side-effect of cleaning up a lot of the code using it.

Finally, I'm using my own (embedded) implementation of std::vector for
both 'children' and 'queue' (std::deque weighs in at 80 bytes on GNU
libstdc++). It's lighter-weight both in memory usage (because it doesn't
include padding) and CPU usage (since it can make some assumptions that
the general std::vector can't).

On my system, a long-running (10,000 attempts) traversal has gone down
from ~1.4 GiB to ~600 MiB, a substantial improvement and about twice the
usage of the coroutines.