Minute TravPos Memory Optimizations
Just changing 'int' to 'short' (I really doubt we'll need more than 65k
productions in a non-terminal) and re-ordering the fields so that the
packing is more ideal.

Note that the queue's size remains an 'int' (now 'unsigned int') because
we really do need that many queue items sometimes.

Also, a new memory allocation policy for the queue. I find that
exponential is overkill for our use-case. Instead, it's linear w.r.t.
the number of children.

Is this micro-optimization? Perhaps. But for a couple-minute change, it
decreases the memory usage by another 20 MiB (for that 10,000 attempt
traversal), which will add up when running the tool for a longer period
of time.