Add Support For (Constant *) and (Variable *)
The '(Variable *)' directive easily maps to '*_VARS';
for '(Constant *)' we currently interpret that as '*_CONSTS', which
means we (naively) extract constants from the constraints and use those.

Additionally, 'replaceFapps' was not being called recursively, which
caused a failed sanity check for the single-invocation solver.

Finally, some minor refactoring.

A better method for extracting constants should be possible, but will
require traversal-level changes and thus take a bit more time.