NewTrav: Cache for 'gettrav'
Profiling suggests that we spend most of our time in NewTrav calling
'gettrav'. This introduces a cache for the function, which is then
invalidated on the subsequent call to 'newtrav'. Caching, like in the
coroutine traversal, is unconditional; this might be changed in the
future if it proves to be problematic.

Note that NewTrav is still slower than Coro in Debug builds, because of
the extra calls to 'gettrav' we do as sanity checks.