Add In-Traversal Support For Unique Variables
For certain use-cases (hint: future commits), it's beneficial to be able
to have a special 'variable' which expands to a unique FAPP each time
it's used in the CFG. This adds support for that, with the
'UNIQUE_SORT_VAR' symbol. So e.g. '(+ UNIQUE_INT_VAR UNIQUE_INT_VAR)'
might expand to '(+ Unique-Var-0 Unique-Var-1)'.
Support exists in the user-specified grammar, but I really don't know
why you'd need that.

Note that this isn't supported with the coroutine traversal, as I can't
come up with a decent way to support it (the path-based method for
NewTrav might work).

More importantly, however, this marks the end of my full support of the
coroutine traversal method. It'll remain as the canonical definition of
the various parameters (SFS, ordered, etc.), but I won't be attempting
to add any more advanced features to it because I don't want to waste
the time doing so. With BFS, NewTrav is practically on-par with it
in terms of speed and memory usage, so there's no need for it any more.

On that note, BFS has been set as the default '--trav_priority'.