Initial Support for Z3-Based '(Constant *)'
Previously, we'd interpret '(Constant *)' as being one of a set of
constants extracted from the program, plus some static built-in ones.
This worked in some cases, but not all.

This commit introduces support for using Z3 for this. Instead of solving
the typical problem of !conjoin(constraints) we now solve
Ec1,...,cn.Vx1,...,xn.constraints, i.e. "Does there exist some set of
constants which makes the candidate function definitions true for all
inputs". If Z3 returns 'sat', then we replace each variable 'ci' with
the model Z3 found for it (i.e. the constant). If it returns 'unsat',
then the candidate definitions are wrong and we pick different ones.

Currently, we don't handle an 'unknown' from Z3; future commits will
enable support for custom enumerative (or otherwise) elimination of
the existential quantifier. None of the SyGuS-Comp 2019 benchmarks that
use the syntactic feature return 'unknown' for the problem (even the
many that use BitVectors) so this doesn't seem to be an enormous problem
yet.