Make Unique Variables More Deterministic
We used to return the set of unique variables in an unordered set, which
with ANY_CONST means we generate constants in a potentially different
order every invocation (if we're instantiating the quantifier ourselves)
and probably introduces some non-determinism across compilations and
systems.

Now, we use an ordered set for hopefully more determinism.