Path Hash Collision Checking
Minor change which adds an optional secondary path tracking method (a
string which we concatenate) to debug the hashes (which ostensibly might
at some point produce a collision). Since this has a very small
overhead, I've enabled it by default.

Also, a small bug fix for crashes with useless non-terminals.