Improve Speed of Compaction
When passing 'true' to AeValSolver::getSkolemFunction, the code's
performance would, in the worst case, scale with O(n!), where n is the
number of iterations required to solve the validity of the problem.
While this is fine for small programs, larger ones often take
>10 iterations which was resulting in an explosion of run time.

Now, we (roughly) scale with O(2^n). While this is still not ideal, this
is still a substantial improvement, and the best I can do without
understanding the algorithm better.