Add 'reset' function to AeValSolver
In the compaction code, we end up having to create a lot of temporary
sub-AeValSolvers; now, we can just reuse a single one, calling 'reset'
on it to set it to the new problem. As long as we are careful not to
reset a solver being used by another part of the code, this provides a
decent speed-up for compaction.