In-Grammar-File Traversal Options
Enables the syntax `(assert (trav_* <expr> <option> <propagate>))` in
grammar files, which enables setting traversal options on a per-NT or
per-production (depending on what `<expr>` is) basis. The syntax is the
same as the command line, so e.g. `(assert (trav_type iterm ordered
false))`. Any options set in the grammar file override those set from
the command line.
The `propagate` option is a Bool and determines whether the changed
option will persist to children. In other words, `(assert (trav_*
ANY_INV * true))` is equivalent to setting the option from the command
line.

Additionally, the `either_inorder` syntax has been added for a
short-hand way of specifying an `either` and
`(assert (trav_type <NT> ordered false))`, meaning that productions are
chosen in order, with all candidates from the first exhausted before the
second is chosen.