Detect BitVector Counters
Just like we have INT_VARS_INC and etc., we now have the corresponding
counters for BitVectors. Additionally we have *_BITINC if the variable
is a power of 2 (or 0) and always increases and the similar *_BITDEC.

Also, NTs without definitions are now treated as useless by the same
code that identifies useless productions, which should allow for better
grammar files.