Initial Commit of In-Traversal Simplification
Previously, we'd call the `simplifyArithm` and `simplifyBool` functions
in ExprSimpl.hpp to simplify candidates before we sent them to the
solver. This acted as a filter, avoiding candidates which were known to
be junk. However, it was fairly inefficient since it required
enumerating all of these candidates, even if simplification could have
determined that, e.g., the left-hand side of an AND is false and thus
the right-hand side doesn't need to be enumerated.

This commit adds support for in-traversal simplification, which - in
essence - can perform short-circuiting of Expr's while traversing, which
is more efficient as shown above. Some rules for the simplification for
Bool's, Int's, and BitVector's are included, though they have not
thoroughly been tested.

Note that this support only exists for NewTrav, since I don't really
want to put in the effort to support it for the Coro traversal (and
don't think it's necessary). As said in an earlier commit, I don't
intend to add these advanced features to Coro, just keeping it as the
'canonical' traversal method for testing purposes.