Correct In-Traversal Pruning
A recent commit introduced in-traversal pruning, with the idea being
that we can e.g. short-circuit within the traversal instead of as a
filter after the fact, which reduces the size of the grammar. However,
this was implemented almost completely incorrectly, and thus was pruning
too aggressively in some cases and failing to prune in others. This has
(as far as I know) been fixed, and pruning should be correct now.

Additionally, I increased the number of past candidates which are
remembered by GramEnum, because this commit sometimes increases the
number of iterations because of old candidates being retried. Note that
this does affect some benchmarks, which rely on this behavior: for
these, just use V2.