GramEnum: Unlimited Old Candidate Detection
As my simplification of candidates has gotten better, we're also getting
better at detecting duplicate candidates. Even with a small limit on the
number of old candidates remembered, we're running into ordering issues.
So increasing the size of the cache, while it does make retrying old
lemmas more essential, is just oil on the flame.

The unlimited cache really doesn't take up that much space, and
in principle should improve performance in some cases.

Also, includes an implementation of LRU/LFU/etc. caches. I really
thought that it would be necessary, but it seems that an unlimited cache
works basically equivalently in most cases.