In-Traversal Old Candidate Detection
We used to detect old candiates at the end, as a filter step. This
requires, however, that try all combinations of bad candidates, even if
we could have pruned them earlier.

Now, each non-terminal keeps track of its previously-generated
candidates, and if a duplicate is found we skip it. Of course, we apply
some heuristics to decide when we don't need to keep track of them.
This notably changes the order that candidates are tried in, so this
may not result in a strict increase in performance. Note this is for
NewTrav only, like with the other commits.

This latest traversal is now even better at ignoring old candidates,
thus V2 is a necessity here. `--retry 10` (or so) is necessary to avoid
tanking performance; `--batch 1` is necessary in some cases for some
reason, I'm not sure why.