Ability to Blacklist Certain Productions
Interface isn't great, but is the best I can come up with for now. The
function takes a 'Path', the same from NewTrav. Blacklisting a path will
also (by nature of the implementation) blacklist paths starting with
that path. The intention of including the path is so we can exclude
productions only at certain depths.