Very Basic CFG Pruning
The core of the idea is that if we can establish that
forall y1,y2,x (!cons(x, y1) ^ !cons(x, y2) => !cons(x, y1 + y2))
(for e.g. production like Start + Start), then we can ignore the given
production. This only happens in some cases: we can (and hopefully will)
consider strengthenings of the negation of the constraints (!cons ^
stren) in order to prove this.

This is "very basic" in that it only supports pruning productions at the
top level, for a single single-application function. This is mostly here
to save my place, the next commit will expand this somewhat.