Different Pruning Interface
Previously, I'd allowed for pruning the grammar by add paths to a
blacklist. However, this turns out to not work quite as well as I'd
like, namely having issues with non-top-level productions. Thus, a new
interface, where the GramEnum user provides a function which determines
if a production should be skipped or not, given the context (determined
by path) where the production is appearing.

It's a fairly general interface, which should be sufficient for the more
complex pruning I want to do. It may, however, be insufficient for
determining pruning in cases of mutual recursion, like we have for
max2.sl.