Fix Memory Leak With TravContext Code
Not only was I not checking to see if an entry already existed in
'findNewCtx' (forgot), i.e. was orphaning old pointer, but I also never
removed entries from 'pathToCtx'; now I do, which should hopefully be
implemented correctly.

Also, added an option for it in TravParams, but not exposing it.