Deeper (Simple) CFG Pruning
This is the same conceptually as the previous commit, but now instead of
doing it in SyGuSSolver, we do it inside of the traversal (by passing a
custom pruning function to GramEnum) using that feature just
implemented. This means we can make pruning decisions about productions
that are deeper than just the top level, but still with the same
limitations as the previous commit (i.e. doesn't work in most cases).