Fix "Unable to Find Inv In Gram" On Success
The message was being printed in the wrong location: it was mostly
right, unless the last candidate of the grammar was the correct one, in
which case it would print the message then finish successful.

Luckily, I have a 'Finish' method which was already being called; I now
print it in there.