Initial Implementation Of Java Interface
Work on a Java GUI has progressed enough for us to want to be able to
interact with FreqHorn natively through Java (instead of by capturing
stdout of the command line process). This is the first pass at that,
featuring the CMake infrastructure necessary to build JNI libraries and
Jar files, and conversion from Expr and Grammar to corresponding Java
interfaces. These interfaces support modifying themselves, so that we
can manipulate Exprs and the Grammar in Java; code to convert these back
to native format will come later, as will a proper interface to the rest
of FreqHorn.