Fix Build Errors Under MSVC
Yes, you read that correctly; I now have the project building under Windows, using vcpkg. It's naming Z3 as "libz3.dll" so we fallback to that in FindZ3.cmake. Also some ambiguous overloading in Boost::range::begin in Expr (might just be because of newer Boost), missing #include <deque> in RndLearnerV3 (no idea why that is), and std::vector complaining about not knowing the size of a ParseTree in ParseTree.hpp.