Initial Commit of Pruning Solver
In essence, we enumerate summaries of non-terminals instead of their
full set of functions, which scales much better. Full details to be
included later somewhere.
Use with `--sygus-method prune`.

It only works on single-application (though I think this is trivial to
fix), single-function (harder to fix), single-non-terminal (I have ideas
on how to fix) problems. Most problems even in this will probably produce
a failed assertion, indicating a limitation with the implementation.
Further commits will fix this (I need a commit point).